% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalBinnedData.R
\name{FinalBinnedData}
\alias{FinalBinnedData}
\title{Intermediate Binned Dataset}
\usage{
FinalBinnedData(input.df, target.var.name)
}
\arguments{
\item{input.df}{Input data frame that contains the target variable and predictor variables with no missing values. Predictors can be either categorical or continuous.}

\item{target.var.name}{Name of binary target variable. Target variables should be numeric with only two distinct values (0, 1)}
}
\value{
\item{TransformedData}{A data frame that contains the transformed dataset used in the HybridFS function}
}
\description{
Retrieves the transformed dataset returned from HybridFS function
}
\examples{
BinnedData=FinalBinnedData(input.df=validation,target.var.name="Survived")
}
