% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-HiFIT.R
\name{hifit}
\alias{hifit}
\title{HiFIT: A high-dimensional feature importance test.}
\usage{
hifit(
  X,
  y,
  k_fold = 5,
  n_perm = 100,
  mod_fun = mod_permfit,
  predict_fun = predict_mod_permfit,
  pathway_list = list(),
  active_var = NULL,
  method = c("random_forest", "lasso", "linear", "svm", "xgboost")[1],
  shuffle = NULL,
  utilities = list(),
  M = 3,
  alpha = 0.1,
  prop = 0.9,
  isAll = FALSE,
  ...
)
}
\arguments{
\item{X}{An n by p numerical matrix. Each row corresponds to an observation.}

\item{y}{The numerical response vector.}

\item{k_fold}{K-fold cross-fitting. If not, set k_fold to zero.}

\item{n_perm}{Number of permutations repeated. Default 100.}

\item{mod_fun}{Function to train the model. Default \code{mod_permfit}.}

\item{predict_fun}{Function to predict the response.
Default \code{predict_mod_permfit}.}

\item{pathway_list}{A list of pathways to be jointly tested.}

\item{active_var}{Index of active variables.}

\item{method}{Models, including \code{ensemble_dnnet} for ensemble deep
neural networks, \code{random_forest} for random forests, \code{lasso}
for linear/logistic lasso, \code{linear} for linear/logistic regression,
\code{svm} for svms with Gaussian kernels, and \code{dnnet} with single deep
neural network.}

\item{shuffle}{If shuffle is null, the data will be shuffled for
cross-fitting; if random shuffle is not desired, please provide
a bector of numbers for cross-fitting indices}

\item{utilities}{Utility functions passed to \code{hybrid.corr}.
Default \code{list()}.}

\item{M}{The order of polynomial regression used by HFS. Default 3.}

\item{alpha}{The cutoff of p-value for feature screening. Default 0.1.}

\item{prop}{The max proportion of screened features. Default 0.9.}

\item{isAll}{Adopt stricter screening standard. Default FALSE.}

\item{...}{Additional parameters passed to each method.}
}
\value{
Returns a PermFIT object.
}
\description{
HiFIT: A high-dimensional feature importance test.
}
