% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-HFS.R
\name{hybrid.corr}
\alias{hybrid.corr}
\title{hybridFS: Hybrid Feature Screening for high-dimensional data.}
\usage{
hybrid.corr(
  X,
  y,
  tau = 0.5,
  prop = 0.9,
  isAll = FALSE,
  utilities = list(),
  M = 3
)
}
\arguments{
\item{X}{An n by p numerical matrix. Each row corresponds to an observation.}

\item{y}{The numerical response vector.}

\item{tau}{A numerical cutoff between 0 and 1 for HFS. Default 0.5.}

\item{prop}{The maximal proportion of kept features. Default 0.9.}

\item{isAll}{If set to \code{TRUE}, set to a stricter standard in screening. Default FALSE.}

\item{utilities}{Default \code{list()}. If not specified, HFS will compute HSIC
and R-squared from polynomial regression instead.}

\item{M}{The order of the polynomial regression if \code{utilities} not specified. Default 3.}
}
\value{
Returns list of variable indexes \code{idx} and HFS scores \code{score}.
}
\description{
hybridFS: Hybrid Feature Screening for high-dimensional data.
}
