% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.perf.R
\name{model.perf}
\alias{model.perf}
\title{Displaying the Optimal Model Performance}
\usage{
model.perf(FS)
}
\arguments{
\item{FS}{FS is the final list returned from HybridFS function.It contains a data frame of the selected features with the relative rank and the optimal model performance.}
}
\description{
Displays the performance metrics of the optimal model generated via the HybridFS function
}
\details{
Displays the performance metrics of the optimal model returned from the HybridFS function such as F1 Score, Accuracy, Precision and Recall.Performance metrics of the Validation dataset is also displayed to understand model stability
}
\examples{
FS=HybridFS(input.df=validation,target.var.name="Survived")
model.perf(FS)
}
