% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PermFIT.R
\name{permfit}
\alias{permfit}
\title{PermFIT: A permutation-based feature importance test.}
\usage{
permfit(
  train,
  validate = NULL,
  k_fold = 5,
  n_perm = 100,
  mod_fun = mod_permfit,
  predict_fun = predict_mod_permfit,
  pathway_list = list(),
  active_var = NULL,
  method = c("random_forest", "lasso", "linear", "svm", "xgboost")[1],
  shuffle = NULL,
  ...
)
}
\arguments{
\item{train}{An dnnetInput object.}

\item{validate}{A validation dataset is required when k_fold = 0.}

\item{k_fold}{K-fold cross-fitting. If not, set k_fold to zero.}

\item{n_perm}{Number of permutations repeated.}

\item{mod_fun}{Function to train the model. Default \code{mod_permfit}.}

\item{predict_fun}{Function to predict the response. Default \code{predict_mod_permfit}.}

\item{pathway_list}{A list of pathways to be jointly tested.}

\item{active_var}{Index of active variables.}

\item{method}{Models, including \code{ensemble_dnnet} for ensemble deep
neural networks, \code{random_forest} for random forests, \code{lasso}
for linear/logistic lasso, \code{linear} for linear/logistic regression,
\code{svm} for svms with Gaussian kernels, and \code{dnnet} with single deep
neural network.}

\item{shuffle}{If shuffle is null, the data will be shuffled for
cross-fitting; if random shuffle is not desired, please provide
a bector of numbers for cross-fitting indices}

\item{...}{Additional parameters passed to each method.}
}
\value{
Returns a PermFIT object.
}
\description{
PermFIT: A permutation-based feature importance test.
}
