% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-HFS.R
\name{tune.tau}
\alias{tune.tau}
\title{hybridFS: Hybrid Feature Screening for high-dimensional data.}
\usage{
tune.tau(
  X,
  y,
  object,
  val_idx = NULL,
  tau_set = c(0.5, 0.55, 0.6, 0.65, 0.7),
  val_ratio = 0.2,
  mod_fun = mod_permfit,
  predict_fun = predict_mod_permfit,
  method = c("random_forest", "lasso", "linear", "svm", "xgboost")[1],
  alpha = 0.1,
  prop = 0.95,
  isAll = FALSE,
  ...
)
}
\arguments{
\item{X}{An n by p numerical matrix. Each row corresponds to an observation.}

\item{y}{The numerical response vector.}

\item{object}{An object from \code{hybrid.corr}.}

\item{val_idx}{The index of validation data. Default \code{NULL}.}

\item{tau_set}{Candidate taus in an increasing order.
Default (0.5, 0.55, 0.6, 0.65, 0.7).}

\item{val_ratio}{If \code{val_idx} is not specified, randomly split the ratio
of data to a validation set. Default 0.2.}

\item{mod_fun}{A training function passed to \code{permfit}.
Default \code{mod_permfit}.}

\item{predict_fun}{A prediction function passed to \code{permfit}.
Defau \code{predict_mod_permfit}.}

\item{method}{Models, including \code{random_forest} for random forests,
\code{lasso} for linear/logistic lasso, \code{linear} for
linear/logistic regression,
 \code{svm} for svms with Gaussian kernels.}

\item{alpha}{The cutoff for p-values. Default 0.1.}

\item{prop}{The maximal proportion of kept features. Default 0.9.}

\item{isAll}{If set to \code{TRUE}, set to a stricter standard in screening.
Default FALSE.}

\item{...}{Other parameters passed to \code{permfit}.}
}
\value{
The list of variable indexes \code{idx} and best tau \code{tau.best}.
}
\description{
hybridFS: Hybrid Feature Screening for high-dimensional data.
}
