% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-data.R
\docType{data}
\name{validation}
\alias{validation}
\title{Survival of passengers on the Titanic}
\format{A data frame with columns:
\describe{
 \item{ID}{Unique Passenger ID}
 \item{Pclass}{1st,2nd,3rd,Crew}
 \item{Sex}{Male,Female}
 \item{Name}{Passenger Name}
 \item{SibSp}{# of siblings / spouses aboard the Titanic}
 \item{Parch}{# of parents / children aboard the Titanic}
 \item{Ticket}{Ticket number}
 \item{Fare}{Passenger fare}
 \item{Cabin}{Cabin number}
 \item{Embarked}{Port of Embarkation}
 \item{Survived}{No,Yes}
}}
\source{
\url{https://www.amstat.org/publications/jse/v3n3/datasets.dawson.html}
}
\description{
Survival of passengers on the Titanic
}
